lspc now has symmetry with swmm
